A new style of the library/CLI
Completely removes both the trait `AppSubCommand`, and all the structs in `rustybeer::calculators::*`, leaving only functions in their place, as they're not needed.
This also tells us several functions/subcommands are not being used, and will help affirm that for future updates.

This is mostly a stylistic change, but I see it as one for the better-I can see how one would argue against this, but at the very least, `&self` should be removed from all the methods for the structs in `rustybeer::calculators::*`, and they should be converted into associated functions, so an instance of `rustybeer::calculators::priming::Priming` (for example) isn't required to call the functions that don't need it, in the slightest